debian: allow depending on "devel" packages
The names of the dev packages of the legacy libibmad and libibumad
is libibmad-devel and libibumad-devel, even for the deb packages.

This was fixed in the rdma-core based packages and this the new ibsim
packaging did not care about it, but now we need to build with the
legacy libraries.

Signed-off-by: Tzafrir Cohen <nvidia@cohens.org.il>